requirements.txt
anyio==3.6.2
appnope==0.1.3
asttokens==2.1.0
asyncpg==0.27.0
autopep8==2.0.0
backcall==0.2.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
databases==0.6.2
decorator==5.1.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
executing==1.2.0
Faker==15.3.4
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
ipython==8.6.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
passlib==1.7.4
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.32
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
stack-data==0.6.0
starlette==0.22.0
tomli==2.0.1
traitlets==5.5.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
